# swagger.yaml
openapi: '3.0.3'
info:
  title: Escopo - Trabalho de Conclusão de Curso SENAI
  version: '1.0'
  description: Documentação das rotas da API do projeto Escopo.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Erro:
      type: object
      properties:
        codigo:
          type: integer
        status:
          type: string
        mensagem:
          type: string
        rota:
          type: string
        data_hora:
          type: string
    ErroCampos:
      type: object
      properties:
        codigo:
          type: integer
        status:
          type: string
        erros:
          type: array
          items:
            type: object
            properties:
              campo:
                type: string
              mensagem:
                type: string
        rota:
          type: string
        data_hora:
          type: string
...
    IntegranteProjeto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nivel_acesso_id:
          type: integer
          example: 2
...
    ProjetoList:
      type: object
      properties:
        id:
          type: integer
          example: 1
        titulo:
          type: string
          example: Meu Projeto
        descricao:
          type: string
          example: Descrição do projeto
        foto_usuarios:
          type: array
          items:
            type: string
          example: ['https://example.com/image1.jpg', 'https://example.com/image2.jpg']
    ProjetoParticipante:
      type: object
      properties:
        nome:
          type: string
          example: Ana Livia
        email:
          type: string
          example: ana@email.com
        projeto_id:
          type: integer
          example: 1
        usuario_id:
          type: integer
          example: 1
        foto_perfil:
          type: string
          nullable: true
          example: null
        nivel_acesso:
          type: string
          example: Gerente de Projeto
        nivel_acesso_id:
          type: integer
          example: 1
        usuario_projeto_id:
          type: integer
          example: 1
    ProjetoParticipantePendente:
      type: object
      properties:
        nome:
          type: string
          example: Marcos Santos
        email:
          type: string
          example: marcos@email.com
        projeto_id:
          type: integer
          example: 1
        usuario_id:
          type: integer
          example: 3
        foto_perfil:
          type: string
          nullable: true
          example: null
        convidado_em:
          type: string
          example: '2025-01-01 03:48:00.000000'
        nivel_acesso:
          type: string
          example: Analista de Requisitos
        nivel_acesso_id:
          type: integer
          example: 2
    ProjetoParticipantes:
      type: object
      properties:
        projeto_id:
          type: integer
          example: 1
        participantes:
          type: array
          items:
            $ref: '#/components/schemas/ProjetoParticipante'
        pendentes:
          type: array
          items:
            $ref: '#/components/schemas/ProjetoParticipantePendente'
    ProjetoDetail:
      type: object
      properties:
        id:
          type: integer
          example: 1
        titulo:
          type: string
          example: Meu Projeto
        descricao:
          type: string
          example: Descrição do projeto
        status:
          type: boolean
          example: true
        data_criacao:
          type: string
          format: date-time
          example: '2026-04-14T10:00:00Z'
        ultima_atualizacao:
          type: string
          format: date-time
          example: '2026-04-14T12:00:00Z'
        nome_responsavel:
          type: string
          example: Nathan
        criador_id:
          type: integer
          example: 1
        nivel_acesso_id:
          type: integer
          example: 2
...
    ProjetoRequest:
      type: object
      required:
        - titulo
      properties:
        titulo:
          type: string
          example: Meu Projeto
          minLength: 1
          maxLength: 100
        descricao:
          type: string
          example: Descrição do projeto
        integrantes:
          type: array
          items:
            $ref: '#/components/schemas/IntegranteProjeto'
...
paths:
...
  /api/v1/projeto:
    post:
      tags:
        - Projeto
      summary: Cria um novo projeto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjetoRequest'
      responses:
        '201':
          description: Projeto criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '400':
          description: Requisição mal formulada, ou ID de integrante inválido, ou ID de nível de acesso inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '422':
          description: Campos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroCampos'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /api/v1/projetos:
    get:
      tags:
        - Projeto
      summary: Lista todos os projetos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjetoList'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /api/v1/projeto/{id}:
    get:
      tags:
        - Projeto
      summary: Retorna detalhes de um projeto específico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID do projeto
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjetoDetail'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Projeto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
    put:
      tags:
        - Projeto
      summary: Atualiza um projeto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID do projeto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjetoRequest'
      responses:
        '200':
          description: Projeto atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  titulo:
                    type: string
                    example: Meu Projeto
                  descricao:
                    type: string
                    example: Descrição do projeto
        '400':
          description: Requisição mal formulada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Um ou mais dos campos inseridos para o integrante não foram encontrados, ou o projeto com o ID informado não foi encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '422':
          description: Campos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroCampos'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
    delete:
      tags:
        - Projeto
      summary: Exclui um projeto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID do projeto
      responses:
        '204':
          description: Projeto excluído
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Projeto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
  /api/v1/projeto/{id}/participantes:
    get:
      tags:
        - Projeto
      summary: Lista participantes e convites pendentes de um projeto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID do projeto
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjetoParticipantes'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '403':
          description: Não possui permissão para acessar esse recurso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Projeto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
  /api/v1/projeto/{id}/categoria:
    post:
      tags:
        - Categoria
      summary: Cria uma categoria de documentação em um projeto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID do projeto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - titulo
              properties:
                titulo:
                  type: string
                  example: Requisitos
                  minLength: 1
                  maxLength: 100
      responses:
        '201':
          description: Categoria criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
        '400':
          description: Requisição mal formulada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Projeto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '422':
          description: Campos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroCampos'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /api/v1/projeto/categoria/{id}:
    delete:
      tags:
        - Categoria
      summary: Deleta uma categoria
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID da categoria
      responses:
        '204':
          description: Categoria excluída
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Projeto ou categoria não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...